+++ molo/surveys/tests/test_views.py
@@ -97,7 +97,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.post(molo_survey_page.url, {
             molo_survey_form_field.label.lower().replace(' ', '-'): 'python'
-        })
+        }, follow=True)
 
         self.assertContains(response, molo_survey_page.thank_you_text)
 
@@ -116,11 +116,9 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         self.assertContains(response, molo_survey_page.title)
         self.assertContains(response, molo_survey_page.intro)
         self.assertContains(response, molo_survey_form_field.label)
-
         response = self.client.post(molo_survey_page.url, {
             molo_survey_form_field.label.lower().replace(' ', '-'): 'python'
-        })
-
+        }, follow=True)
         self.assertContains(response, molo_survey_page.thank_you_text)
 
         # for test_multiple_submissions_not_allowed_by_default_anonymous
@@ -163,7 +161,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
             response = self.client.post(molo_survey_page.url, {
                 molo_survey_form_field.label.lower().replace(' ', '-'):
                     'python'
-            })
+            }, follow=True)
 
             self.assertContains(response, molo_survey_page.thank_you_text)
 
@@ -185,7 +183,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.post(molo_survey_page.url, {
             molo_survey_form_field.label.lower().replace(' ', '-'): 'python'
-        })
+        }, follow=True)
         self.assertContains(response, molo_survey_page.thank_you_text)
         self.assertContains(response, 'Results')
         self.assertContains(response, molo_survey_form_field.label)
@@ -228,7 +226,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         response = self.client.post(molo_survey_page.url + '?p=3', {
             extra_molo_survey_form_field.label.lower().replace(' ', '-'):
                 'Steven Seagal ;)'
-        })
+        }, follow=True)
 
         self.assertContains(response, molo_survey_page.thank_you_text)
 
@@ -254,7 +252,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.post(molo_survey_page.url, {
             molo_survey_form_field.label.lower().replace(' ', '-'): 'python'
-        })
+        }, follow=True)
 
         self.assertContains(response, molo_survey_page.thank_you_text)
 
@@ -297,8 +295,8 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         response = self.client.get('/admin/surveys/')
         self.assertContains(
             response,
-            '<h2><a href="/admin/surveys/submissions/19/">'
-            'Test Survey</a></h2>')
+            '<h2><a href="/admin/surveys/submissions/%s/">'
+            'Test Survey</a></h2>' % molo_survey_page.pk)
         user = get_user_model().objects.create_superuser(
             username='superuser2',
             email='superuser2@email.com', password='pass2')
@@ -307,8 +305,8 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         response = self.client2.get(self.site2.root_url + '/admin/surveys/')
         self.assertNotContains(
             response,
-            '<h2><a href="/admin/surveys/submissions/19/">'
-            'Test Survey</a></h2>')
+            '<h2><a href="/admin/surveys/submissions/%s/">'
+            'Test Survey</a></h2>' % molo_survey_page.pk)
 
     def test_no_duplicate_indexes(self):
         self.assertTrue(SurveysIndexPage.objects.child_of(self.main2).exists())