+++ CHANGES.rst
@@ -1,10 +1,22 @@
 CHANGE LOG
 ==========
 
-3.0.0-beta.1
+5.0.1
+-----
+- Bug Fix: Filter by id for site specific surveys
+
+5.0.0
 -----
 - Added merged cms functionality to surveys
 - Only able to see relevant surveys for site in admin and csv
+
+2.3.0
+-----
+- Add a success url after user submit answers to a survey
+
+2.2.2
+-----
+- Create a success page after user submit answers to a survey
 
 2.2.1
 -----